requirements.txt
fastapi==0.103.2
sqlmodel==0.0.8
typer==0.9.0
uvicorn==0.23.2
python-jose==3.3.0
python-dotenv==1.0.0
passlib==1.7.4
fastapi-pagination==0.12.10
pytest==7.4.2
pytest-factoryboy==2.6.0
SQLAlchemy-Utils==0.41.1
httpx==0.25.0
bcrypt==4.0.1
alembic==1.12.0


